state manager and states functionality added for all except UnlockingState
all chest view/chestStateManager GO's are instantiated in LockedState.
OnButtonClick it will open LockedChestPanel. flow of functions/actions is from chestManager -> ChestSlotService -> UIManager -> LockedChestPanel
with current gems can unlock and it will change state to unlockedState. using action defined in UIManager and subscribed in ChestSlotService, it will run the function which will search for the chest using the index value and from its chestView call the chestStateManager.SwitchState function.

after clicking again on the chest which is now in unlocked state, it will send the data back to chestSlotService of parameter type ChestUnlockData instantiated from the chest model and switch state to collected state which in the current implementation is not being used as the chest is destroyed and replaced with empty chest in the OnChestUnlock function itself.